requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
litellm>=1.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
composio>=0.7.0
google-genai>=1.0.0
chromadb>=1.0.0
numpy>=1.24.0
python-multipart>=0.0.9
langchain>=0.3.0
langchain-community>=0.3.0
unstructured[pdf]>=0.16.0
unstructured-inference>=0.7.0
pytesseract>=0.3.0
pdf2image>=1.17.0
